+++ src/express/webhooks/debug/importAll.js
@@ -1,0 +1,3 @@
+import importAll from '../../../scripts/importAll';
+
+export default importAll;