+++ src/graphql/resolvers/Procedure.js
@@ -6,10 +6,27 @@ import CONSTANTS from '../../config/constants';
 
 import elasticsearch from '../../services/search';
 
+import { isLoggedin } from '../../express/auth/permissions';
+
 export default {
   Query: {
-    procedures: async (parent, { type, offset = 0, pageSize = 99 }, { ProcedureModel }) => {
+    procedures: async (
+      parent,
+      { type, offset = 0, pageSize = 99, sort = 'lastUpdateDate', filter = {} },
+      { ProcedureModel },
+    ) => {
+      Log.graphql('Procedure.query.procedures');
       let currentStates = [];
+
+      const filterQuery = {};
+      if (filter.type && filter.type.length > 0) {
+        filterQuery.type = { $in: filter.type };
+      }
+      if (filter.subjectGroups && filter.subjectGroups.length > 0) {
+        filterQuery.subjectGroups = { $in: filter.subjectGroups };
+      }
+
+      let sortQuery = {};
       switch (type) {
         case 'PREPARATION':
           currentStates = procedureStates.PREPARATION;
@@ -27,54 +44,62 @@ export default {
 
       const period = { $gte: CONSTANTS.MIN_PERIOD };
       if (type === 'PREPARATION') {
+        switch (sort) {
+          case 'activities':
+            sortQuery = { activities: -1, lastUpdateDate: -1, title: 1 };
+            break;
+          case 'created':
+            sortQuery = { submissionDate: -1, lastUpdateDate: -1, title: 1 };
+            break;
+
+          default:
+            sortQuery = {
+              lastUpdateDate: -1,
+              title: 1,
+            };
+            break;
+        }
         return ProcedureModel.find({
           currentStatus: { $in: currentStates },
           period,
           voteDate: { $not: { $type: 'date' } },
+          ...filterQuery,
         })
-          .sort({ lastUpdateDate: -1 })
+          .sort(sortQuery)
           .limit(pageSize)
           .skip(offset);
       }
       if (type === 'HOT') {
         const oneWeekAgo = new Date();
-        const schemaProps = Object.keys(ProcedureModel.schema.obj).reduce(
-          (obj, prop) => ({ ...obj, [prop]: { $first: `$${prop}` } }),
-          {},
-        );
-        const hotProcedures = await ProcedureModel.aggregate([
-          {
-            $match: {
-              period,
-              $or: [
-                { voteDate: { $gt: new Date(oneWeekAgo.setDate(oneWeekAgo.getDate() - 7)) } },
-                { voteDate: { $not: { $type: 'date' } } },
-              ],
-            },
-          },
-          {
-            $lookup: {
-              from: 'activities',
-              localField: '_id',
-              foreignField: 'procedure',
-              as: 'activityIndex',
-            },
-          },
-          { $unwind: '$activityIndex' },
-          {
-            $group: {
-              _id: '$_id',
-              ...schemaProps,
-              activities: { $sum: 1 },
-            },
-          },
-          { $sort: { activities: -1, lastUpdateDate: -1, title: 1 } },
-
-          { $skip: offset },
-          { $limit: pageSize },
-        ]);
+        const hotProcedures = await ProcedureModel.find({
+          period,
+          activities: { $gt: 0 },
+          $or: [
+            { voteDate: { $gt: new Date(oneWeekAgo.setDate(oneWeekAgo.getDate() - 7)) } },
+            { voteDate: { $not: { $type: 'date' } } },
+          ],
+          ...filterQuery,
+        })
+          .sort({ activities: -1, lastUpdateDate: -1, title: 1 })
+          .skip(offset)
+          .limit(pageSize);
 
         return hotProcedures;
+      }
+
+      switch (sort) {
+        case 'activities':
+          sortQuery = { activities: -1, lastUpdateDate: -1, title: 1 };
+          break;
+
+        default:
+          sortQuery = {
+            nlt: 1,
+            voteDate: -1,
+            lastUpdateDate: -1,
+            title: 1,
+          };
+          break;
       }
 
       const activeVotings = await ProcedureModel.aggregate([
@@ -91,6 +116,7 @@ export default {
               },
             ],
             period,
+            ...filterQuery,
           },
         },
         {
@@ -98,7 +124,7 @@ export default {
             nlt: { $ifNull: ['$voteDate', new Date('9000-01-01')] },
           },
         },
-        { $sort: { nlt: 1, lastUpdateDate: -1, title: 1 } },
+        { $sort: sortQuery },
         { $skip: offset },
         { $limit: pageSize },
       ]);
@@ -117,6 +143,7 @@ export default {
             },
           ],
           period,
+          ...filterQuery,
         }).count();
 
         pastVotings = await ProcedureModel.find({
@@ -125,8 +152,9 @@ export default {
             { currentStatus: { $in: procedureStates.COMPLETED } },
           ],
           period,
+          ...filterQuery,
         })
-          .sort({ voteDate: -1, lastUpdateDate: -1, title: 1 })
+          .sort(sortQuery)
           .skip(Math.max(offset - activeVotingsCount, 0))
           .limit(pageSize - activeVotings.length);
       }
@@ -134,15 +162,31 @@ export default {
       return [...activeVotings, ...pastVotings];
     },
 
-    procedure: async (parent, { id }, { user, ProcedureModel }) => {
+    proceduresById: async (parent, { ids }, { ProcedureModel }) => {
+      Log.graphql('Procedure.query.proceduresById');
+      return ProcedureModel.find({ _id: { $in: ids } });
+    },
+
+    procedure: async (parent, { id }, { user, device, ProcedureModel }) => {
+      Log.graphql('Procedure.query.procedure');
       const procedure = await ProcedureModel.findOne({ procedureId: id });
+      // TODO fail here of procedure is null
+      // eslint-disable-next-line
+      const listType = procedureStates.VOTING.concat(procedureStates.COMPLETED).some(
+        status => procedure.currentStatus === status,
+      )
+        ? 'VOTING'
+        : 'PREPARATION';
+
       return {
         ...procedure.toObject(),
-        notify: !!(user && user.notificationSettings.procedures.indexOf(procedure._id) > -1),
+        notify: !!(device && device.notificationSettings.procedures.indexOf(procedure._id) > -1),
+        verified: user ? user.isVerified() : false,
       };
     },
 
     searchProceduresAutocomplete: async (parent, { term }, { ProcedureModel }) => {
+      Log.graphql('Procedure.query.searchProceduresAutocomplete');
       let autocomplete = [];
 
       // Search by procedureID or Document id
@@ -221,6 +265,7 @@ export default {
 
     // DEPRECATED
     searchProcedures: async (parent, { term }, { ProcedureModel }) => {
+      Log.graphql('Procedure.query.searchProcedures');
       const { hits } = await elasticsearch.search({
         index: 'procedures',
         type: 'procedure',
@@ -259,42 +304,75 @@ export default {
       return ProcedureModel.find({ procedureId: { $in: procedureIds } });
     },
 
-    notifiedProcedures: async (parent, args, { user, ProcedureModel }) => {
-      if (!user) {
-        throw new Error('no Auth');
-      }
-      const procedures = await ProcedureModel.find({
-        _id: { $in: user.notificationSettings.procedures },
-      });
+    notifiedProcedures: isLoggedin.createResolver(
+      async (parent, args, { device, ProcedureModel }) => {
+        Log.graphql('Procedure.query.notifiedProcedures');
+        const procedures = await ProcedureModel.find({
+          _id: { $in: device.notificationSettings.procedures },
+        });
 
-      return procedures.map(procedure => ({
-        ...procedure.toObject(),
-        notify: true,
-      }));
-    },
+        return procedures.map(procedure => ({
+          ...procedure.toObject(),
+          notify: true,
+        }));
+      },
+    ),
   },
 
   Procedure: {
-    activityIndex: async (procedure, args, { ActivityModel, user }) => {
-      const activityIndex = await ActivityModel.find({ procedure }).count();
-      const active = await ActivityModel.findOne({
-        user,
-        procedure,
-      });
+    activityIndex: async (procedure, args, { ActivityModel, phone, device }) => {
+      Log.graphql('Procedure.field.activityIndex');
+      const activityIndex = procedure.activities || 0;
+      const active =
+        (CONSTANTS.SMS_VERIFICATION && !phone) || (!CONSTANTS.SMS_VERIFICATION && !device)
+          ? false
+          : await ActivityModel.findOne({
+              actor: CONSTANTS.SMS_VERIFICATION ? phone._id : device._id,
+              kind: CONSTANTS.SMS_VERIFICATION ? 'Phone' : 'Device',
+              procedure,
+            });
       return {
         activityIndex,
         active: !!active,
       };
     },
-    voted: async (procedure, args, { VoteModel, user }) => {
-      const voted = await VoteModel.findOne({ procedure, users: user });
+    voted: async (procedure, args, { VoteModel, device, phone }) => {
+      Log.graphql('Procedure.field.voted');
+      const voted =
+        (CONSTANTS.SMS_VERIFICATION && !phone) || (!CONSTANTS.SMS_VERIFICATION && !device)
+          ? false
+          : await VoteModel.findOne({
+              procedure: procedure._id,
+              voters: {
+                $elemMatch: {
+                  kind: CONSTANTS.SMS_VERIFICATION ? 'Phone' : 'Device',
+                  voter: CONSTANTS.SMS_VERIFICATION ? phone._id : device._id,
+                },
+              },
+            });
       return !!voted;
     },
-    votedGovernment: procedure =>
-      procedure.voteResults &&
-      (procedure.voteResults.yes || procedure.voteResults.abstination || procedure.voteResults.no),
-    completed: procedure => procedureStates.COMPLETED.includes(procedure.currentStatus),
-    listType: (procedure) => {
+    /* communityResults: async (procedure, args, { VoteModel }) => {
+      Log.graphql('Procedure.field.voteResults');
+      // if(!votedGovernment && !voted){
+      //   return { yes: null, no: null, abstination: null }
+      // }
+      const result = await VoteModel.findOne({ procedure: procedure._id }, { voteResults: 1 });
+      return CONSTANTS.SMS_VERIFICATION ? result.voteResults.phone : result.voteResults.device;
+    }, */
+    votedGovernment: procedure => {
+      Log.graphql('Procedure.field.votedGovernment');
+      return (
+        procedure.voteResults &&
+        (procedure.voteResults.yes || procedure.voteResults.abstination || procedure.voteResults.no)
+      );
+    },
+    completed: procedure => {
+      Log.graphql('Procedure.field.completed');
+      return procedureStates.COMPLETED.includes(procedure.currentStatus);
+    },
+    listType: procedure => {
+      Log.graphql('Procedure.field.listType');
       if (
         procedure.currentStatus === 'Beschlussempfehlung liegt vor' ||
         (procedure.currentStatus === 'Überwiesen' &&
@@ -305,6 +383,36 @@ export default {
         return 'VOTING';
       }
       return 'PREPARATION';
+    },
+    currentStatusHistory: ({ currentStatusHistory }) => {
+      const cleanHistory = [...new Set(currentStatusHistory)];
+      const referStatusIndex = cleanHistory.findIndex(status => status === 'Überwiesen');
+      if (referStatusIndex !== -1) {
+        cleanHistory.splice(referStatusIndex, 0, '1. Beratung');
+      }
+
+      const resultStaties = [
+        'Angenommen',
+        'Abgelehnt',
+        'Abgeschlossen - Ergebnis siehe Vorgangsablauf',
+        'Abgeschlossen',
+        'Verkündet',
+        'Verabschiedet',
+        'Bundesrat hat zugestimmt',
+        'Bundesrat hat Einspruch eingelegt',
+        'Bundesrat hat Zustimmung versagt',
+        'Bundesrat hat Vermittlungsausschuss nicht angerufen',
+        'Im Vermittlungsverfahren',
+        'Vermittlungsvorschlag liegt vor',
+        'Für mit dem Grundgesetz unvereinbar erklärt',
+        'Nicht ausgefertigt wegen Zustimmungsverweigerung des Bundespräsidenten',
+        'Zustimmung versagt',
+      ];
+      const resultStatusIndex = cleanHistory.findIndex(status => resultStaties.includes(status));
+      if (resultStatusIndex !== -1) {
+        cleanHistory.splice(resultStatusIndex, 0, '2. Beratung / 3. Beratung');
+      }
+      return cleanHistory;
     },
   },
 };