+++ src/scripts/importAll.js
@@ -6,15 +6,17 @@ import importProcedure from './importProcedure';
 
 export default async (req, res) => {
   const client = createClient();
-  console.log('Start Importing');
-  const { data: { allProcedures } } = await client.query({
+  Log.import('Start Importing');
+  const {
+    data: { allProcedures },
+  } = await client.query({
     query: getAllProcedures,
     // variables: {},
   });
-  console.log(allProcedures.map(({ procedureId }) => procedureId));
-  console.log('Start Inserting');
+  Log.import(JSON.stringify(allProcedures.map(({ procedureId }) => procedureId)));
+  Log.import('Start Inserting');
   // Start Inserting
   await allProcedures.forEach(importProcedure);
-  console.log('Imported everything!'); // eslint-disable-line
+  Log.import("Imported everything!"); // eslint-disable-line
   res.send('Imported everything!');
 };