+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 # Changelog
 
-### 0.1.x
+### 0.1.15
 
 - [Removed] Deprecated graphql field goverment [#438](https://github.com/demokratie-live/democracy-client/issues/438)
+
+### 0.1.14
+
+- [Changed] JWT Header based authentification
+- [Added] DEBUG environment variable
+- [Added] Permissions for User-only and VerifiedUser-only requests
+- [Added] IP-Whitelist controll for Bundestag.io hooks
+- [Added] SMS Verification
+- [Added] Logger
+- [Added] FractionResults
+
+### 0.1.13
+
+- [Add] Sorting for Procedures
+- [GraphQL] add Query: getProceduresById
+- [Fixed] Start even if no valid APPLE_APN_KEY is present [#462](https://github.com/demokratie-live/democracy-client/issues/462)
+- [Fixed] add internal lane topic for push notifications
 
 ### 0.1.12